Move user query to `db/users.rs`
